Refactor location of encode decode value functions
