+++ src/kompressor/utils.py
@@ -21,153 +21,6 @@
 # SOFTWARE.
 
 
-import jax
-import jax.numpy as jnp
-
-
-@jax.jit
-def encode_values_raw(pred, gt):
-    return jnp.int32(gt) - jnp.int32(pred)
-
-
-@jax.jit
-def decode_values_raw(pred, encoded):
-    return jnp.int32(pred) + jnp.int32(encoded)
-
-
-@jax.jit
-def encode_values_uint8(pred, gt):
-    # Encode predictions in [0, +256) to deltas in [-256, +256) to encoded values in [0, +256)
-    # Leads to peaks at 0 and +255
-    return jnp.uint8(((jnp.int32(gt) - jnp.int32(pred)) + 256) % 256)
-
-@jax.jit
-def decode_values_uint8(pred, encoded):
-    # Decode twin peaks distribution
-    return jnp.uint8(((jnp.int32(pred) + jnp.int32(encoded)) + 256) % 256)
-
-
-@jax.jit
-def encode_transform_interleaved_uint8(input):
-    # Take twin peaks at 0 and +255 and shift them to be a single peak centered around 0 in [-128, +128)
-    # Encode positive values a strided by 1, and negatives values interleaved between them in [0, +256)
-    encoded = ((jnp.int32(input) + 128) % 256) - 128
-    encoded = jnp.where((encoded >= 0), (encoded * 2), ((encoded + 1) * -2) + 1)
-    return jnp.uint8(encoded)
-
-
-@jax.jit
-def decode_transform_interleaved_uint8(encoded):
-    # Decode interleaved in [0, +256) to [-128, +128) to [0, 256+)
-    # Positive values strided by 1, and negatives values interleaved between them
-    # Leads to peaks at 0 and +255
-    encoded = jnp.int32(encoded)
-    encoded = jnp.where((encoded % 2 == 0), (encoded // 2), -(((encoded - 1) // 2) + 1))
-    return jnp.uint8((encoded + 256) % 256)
-
-
-@jax.jit
-def encode_transform_centre_uint8(input):
-    # Take twin peaks at 0 and +255 and shift them to be a single peak centered around +128 in [0, +256)
-    return jnp.uint8((jnp.int32(input) + 128) % 256)
-
-
-@jax.jit
-def decode_transform_centre_uint8(encoded):
-    # Decode single peak at +128 into twin peaks at 0 and +255 in [0, +256)
-    return jnp.uint8((jnp.int32(encoded) + 128) % 256)
-
-
-@jax.jit
-def encode_values_uint16(pred, gt):
-    return jnp.uint16(((jnp.int32(gt) - jnp.int32(pred)) + 65536) % 65536)
-
-
-@jax.jit
-def decode_values_uint16(pred, encoded):
-    return jnp.uint16(((jnp.int32(pred) + jnp.int32(encoded)) + 65536) % 65536)
-
-
-@jax.jit
-def encode_transform_interleaved_uint16(input):
-    encoded = ((jnp.int32(input) + 32768) % 65536) - 32768
-    encoded = jnp.where((encoded >= 0), (encoded * 2), ((encoded + 1) * -2) + 1)
-    return jnp.uint16(encoded)
-
-
-@jax.jit
-def decode_transform_interleaved_uint16(encoded):
-    encoded = jnp.int32(encoded)
-    encoded = jnp.where((encoded % 2 == 0), (encoded // 2), -(((encoded - 1) // 2) + 1))
-    return jnp.uint16((encoded + 65536) % 65536)
-
-
-@jax.jit
-def encode_transform_centre_uint16(input):
-    return jnp.uint16((jnp.int32(input) + 32768) % 65536)
-
-
-@jax.jit
-def decode_transform_centre_uint16(encoded):
-    return jnp.uint16((jnp.int32(encoded) + 32768) % 65536)
-
-
-@jax.jit
-def encode_categorical(pred, gt):
-
-    # Output shape and dtype determined from reference values
-    shape = gt.shape
-    dtype = gt.dtype
-
-    # Determine the descending order indexing for the logits at each spatial location and channel
-    logit_ranks = jnp.argsort(pred).astype(dtype)[..., ::-1]
-
-    # Flatten logit ranks from [B, ...SPATIAL..., C, L] to [-1, L]
-    flat_logit_ranks = jnp.reshape(logit_ranks, (-1, logit_ranks.shape[-1]))
-    # Flatten gt from [B, ...SPATIAL..., C] to [-1,]
-    flat_gt = jnp.reshape(gt, (-1,))
-
-    # Perform argwhere for a single location and channel
-    def argwhere(logit_ranks, gt):
-        return jnp.argmax(logit_ranks == gt).astype(dtype)
-
-    # Distribute the argwhere over all spatial locations and channels
-    flat_encoded = jax.vmap(argwhere)(flat_logit_ranks, flat_gt)
-
-    # Reshape the encoded values back into the shape of the gt tensor
-    encoded = jnp.reshape(flat_encoded, shape)
-
-    return encoded
-
-
-@jax.jit
-def decode_categorical(pred, encoded):
-
-    # Output shape and dtype determined from encoded values
-    shape = encoded.shape
-    dtype = encoded.dtype
-
-    # Determine the descending order indexing for the logits at each spatial location and channel
-    logit_ranks = jnp.argsort(pred).astype(dtype)[..., ::-1]
-
-    # Flatten logit ranks from [B, ...SPATIAL..., C, L] to [-1, L]
-    flat_logit_ranks = jnp.reshape(logit_ranks, (-1, logit_ranks.shape[-1]))
-    # Flatten encoded from [B, ...SPATIAL..., C] to [-1,]
-    flat_encoded = jnp.reshape(encoded, (-1,))
-
-    # Perform indexing for a single location and channel
-    def index(logit_ranks, encoded):
-        return logit_ranks[encoded]
-
-    # Distribute the indexing over all spatial locations and channels
-    flat_decoded = jax.vmap(index)(flat_logit_ranks, flat_encoded)
-
-    # Reshape the decoded values back into the shape of the encoded tensor
-    decoded = jnp.reshape(flat_decoded, shape)
-
-    return decoded
-
-
 def yield_chunks(max_value, chunk):
     # Assert max value is positive
     assert max_value > 0
